Rename file_handling directory to file
And split file_handling/index.js into general utility methods (readAsArrayBuffer), and rom related methods.
